+++ mvc/src/main/java/nextstep/mvc/controller/asis/InterfaceBasedControllerHandlerAdapter.java
@@ -1,9 +1,8 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller.asis;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.asis.Controller;
 import nextstep.mvc.controller.exception.UnsupportedViewFileExtensionException;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
